Refactor to solve ambiguous prepare/create behaviour.

+++ program/source/window.cpp
@@ -5,6 +5,8 @@
 #include "SDL3/SDL_events.h"
 #include "glm/ext/vector_uint2.hpp"
 
+#include "exception.hpp"
+
 namespace cse
 {
   window::window(const std::string &title_, const glm::uvec2 &dimensions_, const bool fullscreen_, const bool vsync_)
@@ -30,17 +32,37 @@ namespace cse
 
   window::~window() { hook.reset(); }
 
-  void window::initialize()
+  void window::prepare()
   {
+    if (state.prepared) throw exception("Window cannot be prepared more than once");
+    if (state.created) throw exception("Window cannot be prepared while created");
+    state.active.running = true;
+    state.prepared = true;
+    hook.call<void()>("prepare");
+  }
+
+  void window::create()
+  {
+    if (!state.prepared) throw exception("Window must be prepared before creation");
+    if (state.created) throw exception("Window cannot be created more than once");
     graphics.create_window(state.active.width, state.active.height, state.active.left, state.active.top,
                            state.active.display_index, state.active.fullscreen, state.active.vsync);
-    state.active.running = true;
-    state.initialized = true;
-    hook.call<void()>("initialize");
+    state.created = true;
+    hook.call<void()>("create");
+  }
+
+  void window::previous()
+  {
+    if (!state.prepared) throw exception("Window must be prepared before updating previous state");
+    if (!state.created) throw exception("Window must be created before updating previous state");
+    state.update_previous();
+    graphics.update_previous();
   }
 
   void window::event()
   {
+    if (!state.prepared) throw exception("Window must be prepared before processing events");
+    if (!state.created) throw exception("Window must be created before processing events");
     switch (state.event.type)
     {
       case SDL_EVENT_QUIT: state.active.running = false; break;
@@ -55,36 +77,54 @@ namespace cse
     }
   }
 
-  void window::input() { hook.call<void(const bool *)>("input", state.input); }
-
-  void window::simulate(const float poll_rate) { hook.call<void(const float)>("simulate", poll_rate); }
-
-  bool window::start_render(const double alpha, const float aspect_ratio)
+  void window::input()
   {
+    if (!state.prepared) throw exception("Window must be prepared before processing input");
+    if (!state.created) throw exception("Window must be created before processing input");
+    hook.call<void(const bool *)>("input", state.input);
+  }
+
+  void window::simulate(const float poll_rate)
+  {
+    if (!state.prepared) throw exception("Window must be prepared before simulation");
+    if (!state.created) throw exception("Window must be created before simulation");
+    hook.call<void(const float)>("simulate", poll_rate);
+  }
+
+  bool window::pre_render(const double alpha, const float aspect_ratio)
+  {
+    if (!state.prepared) throw exception("Window must be prepared before starting rendering");
+    if (!state.created) throw exception("Window must be created before starting rendering");
     if (!graphics.acquire_swapchain_texture()) return false;
     graphics.start_render_pass(state.active.width, state.active.height, aspect_ratio);
     hook.call<void(const double)>("pre_render", alpha);
     return true;
   }
 
-  void window::end_render(const double alpha)
+  void window::post_render(const double alpha)
   {
+    if (!state.prepared) throw exception("Window must be prepared before ending rendering");
+    if (!state.created) throw exception("Window must be created before ending rendering");
     graphics.end_render_pass();
     hook.call<void(const double)>("post_render", alpha);
   }
 
-  void window::cleanup()
+  void window::destroy()
   {
+    if (!state.prepared) throw exception("Window must be prepared before destruction");
+    if (!state.created) throw exception("Window cannot be destroyed more than once");
     state.input = nullptr;
     state.event = {};
     graphics.destroy_window();
-    state.initialized = false;
-    hook.call<void()>("cleanup");
+    state.created = false;
+    hook.call<void()>("destroy");
   }
 
-  void window::previous()
+  void window::clean()
   {
-    state.update_previous();
-    graphics.update_previous();
+    if (!state.prepared) throw exception("Window cannot be cleaned more than once");
+    if (state.created) throw exception("Window must be destroyed before cleaning");
+    state.prepared = false;
+    hook.call<void()>("clean");
   }
 }
